refactor(maturity): extract the combine merge-sources sub-section (#467)
Second combine sub-section split. Move the merge-sibling-deposits picker out of
MaturityResolveBody into `MergeSourcesSection` (new file): the eligibility window
slider, the per-source list (blocked-with-reason / selectable / overridable-early),
the per-source received inputs + overall total, and the destination bank for a
multi-source merge. The sheet keeps the state (mergeSel/mergeRecv/windowDays/…),
the eligibility classification map, and the toggle/override handlers, and passes
them in. The now-unused Lock / SlidersHorizontal imports drop out of the sheet.
No behavior or markup change.

Parity: the MaturityResolveSheet component tests + the maturity-combine-merge E2E
(settle sibling early, window gating, cross-bank destination, oversized-received
rejection) both green.

Verified: full unit suite 1261 passed; typecheck 0 errors; lint 0 errors; E2E
maturity-combine-merge (5) passed. MaturityResolveSheet 1195 → 1109.
